Module google_jobs4::api

Structs

ApplicationInfo
Application related details of a job posting.

BatchCreateJobsRequest
Request to create a batch of jobs.

BatchDeleteJobsRequest
Request to delete a batch of jobs.

BatchUpdateJobsRequest
Request to update a batch of jobs.

ClientEvent
An event issued when an end user interacts with the application that implements Cloud Talent Solution. Providing this information improves the quality of results for the API clients, enabling the service to perform optimally. The number of events sent must be consistent with other calls, such as job searches, issued to the service by the client.

CloudTalentSolution
Central instance to access all CloudTalentSolution related resource activities

CommuteFilter
Parameters needed for commute search.

CommuteInfo
Commute details related to this job.

Company
A Company resource represents a company in the service. A company is the entity that owns job postings, that is, the hiring entity responsible for employing applicants for the job position.

CompanyDerivedInfo
Derived details about the company.

CompensationEntry
A compensation entry that represents one component of compensation, such as base pay, bonus, or other compensation type. Annualization: One compensation entry can be annualized if - it contains valid amount or range. - and its expected_units_per_year is set or can be derived. Its annualized range is determined as (amount or range) times expected_units_per_year.

CompensationFilter
Filter on job compensation type and amount.

CompensationInfo
Job compensation details.

CompensationRange
Compensation range.

CompleteQueryResponse
Response of auto-complete query.

CompletionResult
Resource that represents completion results.

CustomAttribute
Custom attribute values that are either filterable or non-filterable.

CustomRankingInfo
Custom ranking information for SearchJobsRequest.

DeviceInfo
Device information collected from the job seeker, candidate, or other entity conducting the job search. Providing this information improves the quality of the search results across devices.

Empty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance: service Foo { rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty); } The JSON representation for Empty is empty JSON object {}.

HistogramQuery
The histogram request.

HistogramQueryResult
Histogram result that matches HistogramQuery specified in searches.

Job
A Job resource represents a job posting (also referred to as a "job listing" or "job requisition"). A job belongs to a Company, which is the hiring entity responsible for the job.

JobDerivedInfo
Derived details about the job posting.

JobEvent
An event issued when a job seeker interacts with the application that implements Cloud Talent Solution.

JobQuery
The query required to perform a search query.

LatLng
An object that represents a latitude/longitude pair. This is expressed as a pair of doubles to represent degrees latitude and degrees longitude. Unless specified otherwise, this must conform to the WGS84 standard. Values must be within normalized ranges.

ListCompaniesResponse
The List companies response object.

ListJobsResponse
List jobs response.

ListTenantsResponse
The List tenants response object.

Location
A resource that represents a location with full geographic information.

LocationFilter
Geographic region of the search.

MatchingJob
Job entry with metadata inside SearchJobsResponse.

Money
Represents an amount of money with its currency type.

Operation
This resource represents a long-running operation that is the result of a network API call.

PostalAddress
Represents a postal address, e.g. for postal delivery or payments addresses. Given a postal address, a postal service can deliver items to a premise, P.O. Box or similar. It is not intended to model geographical locations (roads, towns, mountains). In typical usage an address would be created via user input or from importing existing data, depending on the type of process. Advice on address input / editing: - Use an i18n-ready address widget such as https://github.com/google/libaddressinput) - Users should not be presented with UI elements for input or editing of fields outside countries where that field is used. For more guidance on how to use this schema, please see: https://support.google.com/business/answer/6397478

ProcessingOptions
Options for job processing.

ProjectMethods
A builder providing access to all methods supported on project resources. It is not used directly, but through the CloudTalentSolution hub.

ProjectOperationGetCall
Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

ProjectTenantClientEventCreateCall
Report events issued when end user interacts with customer's application that uses Cloud Talent Solution. You may inspect the created events in self service tools. Learn more about self service tools.

ProjectTenantCompanyCreateCall
Creates a new company entity.

ProjectTenantCompanyDeleteCall
Deletes specified company. Prerequisite: The company has no jobs associated with it.

ProjectTenantCompanyGetCall
Retrieves specified company.

ProjectTenantCompanyListCall
Lists all companies associated with the project.

ProjectTenantCompanyPatchCall
Updates specified company.

ProjectTenantCompleteQueryCall
Completes the specified prefix with keyword suggestions. Intended for use by a job search auto-complete search box.

ProjectTenantCreateCall
Creates a new tenant entity.

ProjectTenantDeleteCall
Deletes specified tenant.

ProjectTenantGetCall
Retrieves specified tenant.

ProjectTenantJobBatchCreateCall
Begins executing a batch create jobs operation.

ProjectTenantJobBatchDeleteCall
Begins executing a batch delete jobs operation.

ProjectTenantJobBatchUpdateCall
Begins executing a batch update jobs operation.

ProjectTenantJobCreateCall
Creates a new job. Typically, the job becomes searchable within 10 seconds, but it may take up to 5 minutes.

ProjectTenantJobDeleteCall
Deletes the specified job. Typically, the job becomes unsearchable within 10 seconds, but it may take up to 5 minutes.

ProjectTenantJobGetCall
Retrieves the specified job, whose status is OPEN or recently EXPIRED within the last 90 days.

ProjectTenantJobListCall
Lists jobs by filter.

ProjectTenantJobPatchCall
Updates specified job. Typically, updated contents become visible in search results within 10 seconds, but it may take up to 5 minutes.

ProjectTenantJobSearchCall
Searches for jobs using the provided SearchJobsRequest. This call constrains the visibility of jobs present in the database, and only returns jobs that the caller has permission to search against.

ProjectTenantJobSearchForAlertCall
Searches for jobs using the provided SearchJobsRequest. This API call is intended for the use case of targeting passive job seekers (for example, job seekers who have signed up to receive email alerts about potential job opportunities), it has different algorithmic adjustments that are designed to specifically target passive job seekers. This call constrains the visibility of jobs present in the database, and only returns jobs the caller has permission to search against.

ProjectTenantListCall
Lists all tenants associated with the project.

ProjectTenantPatchCall
Updates specified tenant.

RequestMetadata
Meta information related to the job searcher or entity conducting the job search. This information is used to improve the performance of the service.

ResponseMetadata
Additional information returned to client, such as debugging information.

SearchJobsRequest
The Request body of the SearchJobs call.

SearchJobsResponse
Response for SearchJob method.

SpellingCorrection
Spell check result.

Status
The Status type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by gRPC. Each Status message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the API Design Guide.

Tenant
A Tenant resource represents a tenant in the service. A tenant is a group or entity that shares common access with specific privileges for resources like jobs. Customer may create multiple tenants to provide data isolation for different groups.

TimeOfDay
Represents a time of day. The date and time zone are either not significant or are specified elsewhere. An API may choose to allow leap seconds. Related types are google.type.Date and google.protobuf.Timestamp.

TimestampRange
Message representing a period of time between two timestamps.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.