Struct google_jobs4::api::ProjectMethods

pub struct ProjectMethods<'a, C> where
    C: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on project resources. It is not used directly, but through the CloudTalentSolution hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_jobs4 as jobs4;
 
use std::default::Default;
use oauth2;
use jobs4::CloudTalentSolution;
 
let secret: oauth2::ApplicationSecret = Default::default();
let auth = yup_oauth2::InstalledFlowAuthenticator::builder(
        secret,
        yup_oauth2::InstalledFlowReturnMethod::HTTPRedirect,
    ).build().await.unwrap();
let mut hub = CloudTalentSolution::new(hyper::Client::builder().build(hyper_rustls::HttpsConnector::with_native_roots()), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `operations_get(...)`, `tenants_client_events_create(...)`, `tenants_companies_create(...)`, `tenants_companies_delete(...)`, `tenants_companies_get(...)`, `tenants_companies_list(...)`, `tenants_companies_patch(...)`, `tenants_complete_query(...)`, `tenants_create(...)`, `tenants_delete(...)`, `tenants_get(...)`, `tenants_jobs_batch_create(...)`, `tenants_jobs_batch_delete(...)`, `tenants_jobs_batch_update(...)`, `tenants_jobs_create(...)`, `tenants_jobs_delete(...)`, `tenants_jobs_get(...)`, `tenants_jobs_list(...)`, `tenants_jobs_patch(...)`, `tenants_jobs_search(...)`, `tenants_jobs_search_for_alert(...)`, `tenants_list(...)` and `tenants_patch(...)`
// to build up your call.
let rb = hub.projects();

Implementations

impl<'a, C> ProjectMethods<'a, C>

pub fn operations_get(&self, name: &str) -> ProjectOperationGetCall<'a, C>

Create a builder to help you perform the following task:

Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

Arguments

name - The name of the operation resource.

pub fn tenants_client_events_create(
&self,
request: ClientEvent,
parent: &str
) -> ProjectTenantClientEventCreateCall<'a, C>

Create a builder to help you perform the following task:

Report events issued when end user interacts with customer's application that uses Cloud Talent Solution. You may inspect the created events in self service tools. Learn more about self service tools.

Arguments

request - No description provided.

parent - Required. Resource name of the tenant under which the event is created. The format is "projects/{project_id}/tenants/{tenant_id}", for example, "projects/foo/tenants/bar".

pub fn tenants_companies_create(
&self,
request: Company,
parent: &str
) -> ProjectTenantCompanyCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a new company entity.

Arguments

request - No description provided.

parent - Required. Resource name of the tenant under which the company is created. The format is "projects/{project_id}/tenants/{tenant_id}", for example, "projects/foo/tenants/bar".

pub fn tenants_companies_delete(
&self,
name: &str
) -> ProjectTenantCompanyDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes specified company. Prerequisite: The company has no jobs associated with it.

Arguments

name - Required. The resource name of the company to be deleted. The format is "projects/{project_id}/tenants/{tenant_id}/companies/{company_id}", for example, "projects/foo/tenants/bar/companies/baz".

pub fn tenants_companies_get(
&self,
name: &str
) -> ProjectTenantCompanyGetCall<'a, C>

Create a builder to help you perform the following task:

Retrieves specified company.

Arguments

name - Required. The resource name of the company to be retrieved. The format is "projects/{project_id}/tenants/{tenant_id}/companies/{company_id}", for example, "projects/api-test-project/tenants/foo/companies/bar".

pub fn tenants_companies_list(
&self,
parent: &str
) -> ProjectTenantCompanyListCall<'a, C>

Create a builder to help you perform the following task:

Lists all companies associated with the project.

Arguments

parent - Required. Resource name of the tenant under which the company is created. The format is "projects/{project_id}/tenants/{tenant_id}", for example, "projects/foo/tenants/bar".

pub fn tenants_companies_patch(
&self,
request: Company,
name: &str
) -> ProjectTenantCompanyPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates specified company.

Arguments

request - No description provided.

name - Required during company update. The resource name for a company. This is generated by the service when a company is created. The format is "projects/{project_id}/tenants/{tenant_id}/companies/{company_id}", for example, "projects/foo/tenants/bar/companies/baz".

pub fn tenants_jobs_batch_create(
&self,
request: BatchCreateJobsRequest,
parent: &str
) -> ProjectTenantJobBatchCreateCall<'a, C>

Create a builder to help you perform the following task:

Begins executing a batch create jobs operation.

Arguments

request - No description provided.

parent - Required. The resource name of the tenant under which the job is created. The format is "projects/{project_id}/tenants/{tenant_id}". For example, "projects/foo/tenants/bar".

pub fn tenants_jobs_batch_delete(
&self,
request: BatchDeleteJobsRequest,
parent: &str
) -> ProjectTenantJobBatchDeleteCall<'a, C>

Create a builder to help you perform the following task:

Begins executing a batch delete jobs operation.

Arguments

request - No description provided.

parent - Required. The resource name of the tenant under which the job is created. The format is "projects/{project_id}/tenants/{tenant_id}". For example, "projects/foo/tenants/bar". The parent of all of the jobs specified in names must match this field.

pub fn tenants_jobs_batch_update(
&self,
request: BatchUpdateJobsRequest,
parent: &str
) -> ProjectTenantJobBatchUpdateCall<'a, C>

Create a builder to help you perform the following task:

Begins executing a batch update jobs operation.

Arguments

request - No description provided.

parent - Required. The resource name of the tenant under which the job is created. The format is "projects/{project_id}/tenants/{tenant_id}". For example, "projects/foo/tenants/bar".

pub fn tenants_jobs_create(
&self,
request: Job,
parent: &str
) -> ProjectTenantJobCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a new job. Typically, the job becomes searchable within 10 seconds, but it may take up to 5 minutes.

Arguments

request - No description provided.

parent - Required. The resource name of the tenant under which the job is created. The format is "projects/{project_id}/tenants/{tenant_id}". For example, "projects/foo/tenants/bar".

pub fn tenants_jobs_delete(
&self,
name: &str
) -> ProjectTenantJobDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes the specified job. Typically, the job becomes unsearchable within 10 seconds, but it may take up to 5 minutes.

Arguments

name - Required. The resource name of the job to be deleted. The format is "projects/{project_id}/tenants/{tenant_id}/jobs/{job_id}". For example, "projects/foo/tenants/bar/jobs/baz".

pub fn tenants_jobs_get(&self, name: &str) -> ProjectTenantJobGetCall<'a, C>

Create a builder to help you perform the following task:

Retrieves the specified job, whose status is OPEN or recently EXPIRED within the last 90 days.

Arguments

name - Required. The resource name of the job to retrieve. The format is "projects/{project_id}/tenants/{tenant_id}/jobs/{job_id}". For example, "projects/foo/tenants/bar/jobs/baz".

pub fn tenants_jobs_list(&self, parent: &str) -> ProjectTenantJobListCall<'a, C>

Create a builder to help you perform the following task:

Lists jobs by filter.

Arguments

parent - Required. The resource name of the tenant under which the job is created. The format is "projects/{project_id}/tenants/{tenant_id}". For example, "projects/foo/tenants/bar".

pub fn tenants_jobs_patch(
&self,
request: Job,
name: &str
) -> ProjectTenantJobPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates specified job. Typically, updated contents become visible in search results within 10 seconds, but it may take up to 5 minutes.

Arguments

request - No description provided.

name - Required during job update. The resource name for the job. This is generated by the service when a job is created. The format is "projects/{project_id}/tenants/{tenant_id}/jobs/{job_id}". For example, "projects/foo/tenants/bar/jobs/baz". Use of this field in job queries and API calls is preferred over the use of requisition_id since this value is unique.

pub fn tenants_jobs_search(
&self,
request: SearchJobsRequest,
parent: &str
) -> ProjectTenantJobSearchCall<'a, C>

Create a builder to help you perform the following task:

Searches for jobs using the provided SearchJobsRequest. This call constrains the visibility of jobs present in the database, and only returns jobs that the caller has permission to search against.

Arguments

request - No description provided.

parent - Required. The resource name of the tenant to search within. The format is "projects/{project_id}/tenants/{tenant_id}". For example, "projects/foo/tenants/bar".

pub fn tenants_jobs_search_for_alert(
&self,
request: SearchJobsRequest,
parent: &str
) -> ProjectTenantJobSearchForAlertCall<'a, C>

Create a builder to help you perform the following task:

Searches for jobs using the provided SearchJobsRequest. This API call is intended for the use case of targeting passive job seekers (for example, job seekers who have signed up to receive email alerts about potential job opportunities), it has different algorithmic adjustments that are designed to specifically target passive job seekers. This call constrains the visibility of jobs present in the database, and only returns jobs the caller has permission to search against.

Arguments

request - No description provided.

parent - Required. The resource name of the tenant to search within. The format is "projects/{project_id}/tenants/{tenant_id}". For example, "projects/foo/tenants/bar".

pub fn tenants_complete_query(
&self,
tenant: &str
) -> ProjectTenantCompleteQueryCall<'a, C>

Create a builder to help you perform the following task:

Completes the specified prefix with keyword suggestions. Intended for use by a job search auto-complete search box.

Arguments

tenant - Required. Resource name of tenant the completion is performed within. The format is "projects/{project_id}/tenants/{tenant_id}", for example, "projects/foo/tenants/bar".

pub fn tenants_create(
&self,
request: Tenant,
parent: &str
) -> ProjectTenantCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a new tenant entity.

Arguments

request - No description provided.

parent - Required. Resource name of the project under which the tenant is created. The format is "projects/{project_id}", for example, "projects/foo".

pub fn tenants_delete(&self, name: &str) -> ProjectTenantDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes specified tenant.

Arguments

name - Required. The resource name of the tenant to be deleted. The format is "projects/{project_id}/tenants/{tenant_id}", for example, "projects/foo/tenants/bar".

pub fn tenants_get(&self, name: &str) -> ProjectTenantGetCall<'a, C>

Create a builder to help you perform the following task:

Retrieves specified tenant.

Arguments

name - Required. The resource name of the tenant to be retrieved. The format is "projects/{project_id}/tenants/{tenant_id}", for example, "projects/foo/tenants/bar".

pub fn tenants_list(&self, parent: &str) -> ProjectTenantListCall<'a, C>

Create a builder to help you perform the following task:

Lists all tenants associated with the project.

Arguments

parent - Required. Resource name of the project under which the tenant is created. The format is "projects/{project_id}", for example, "projects/foo".

pub fn tenants_patch(
&self,
request: Tenant,
name: &str
) -> ProjectTenantPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates specified tenant.

Arguments

request - No description provided.

name - Required during tenant update. The resource name for a tenant. This is generated by the service when a tenant is created. The format is "projects/{project_id}/tenants/{tenant_id}", for example, "projects/foo/tenants/bar".